fix buttons, add default negative to env, tidy
buttons were not grabbing the right image info before. also note that message history permission must be on for the active channel for buttons to work!

interrogate doesn't happen every button press now. the prompt is retained (except the style prompt which is updated if the style changes from pressing the ReStyle button)